README.md
The Direct aims for quick, accurate and complete reporting at all times. If we ever need to make a correction to our work, we will do so as swiftly as we are made aware of the error. The Direct's goal is to get accurate information to our readers as fast as possible, so any breaking news stories will be expanded on after publication of the original article if necessary.